# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Starships Project OpenApi",
        "description": "Starships Project OpenApi",
        "version": "1.0.0"
    },
    "paths": {
        "/api/people": {
            "get": {
                "tags": [
                    "people"
                ],
                "summary": "Get list of people",
                "description": "Returns list of people",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "people"
                ],
                "summary": "Store a person",
                "description": "Stores a person",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Person"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/people/{id}": {
            "get": {
                "tags": [
                    "people"
                ],
                "summary": "Get person",
                "description": "Returns a person",
                "operationId": "get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "put": {
                "tags": [
                    "people"
                ],
                "summary": "Update a person",
                "description": "Update a person",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Person"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "people"
                ],
                "summary": "Delete person",
                "description": "Delete a person",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Person": {
                "title": "Person model",
                "description": "Person model",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "name"
                    },
                    "mass": {
                        "title": "Mass",
                        "description": "mass"
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}